+++ controllers/ChecklistController.php
@@ -8,6 +8,7 @@ use humhub\modules\content\components\ContentContainerControllerAccess;
 use humhub\modules\space\models\Space;
 use humhub\modules\tasks\models\checklist\CheckForm;
 use humhub\modules\tasks\permissions\ManageTasks;
+use humhub\modules\user\models\User;
 use Yii;
 
 class ChecklistController extends AbstractTaskController
@@ -15,7 +16,7 @@ class ChecklistController extends AbstractTaskController
     public function getAccessRules()
     {
         return [
-            [ContentContainerControllerAccess::RULE_USER_GROUP_ONLY => [Space::USERGROUP_MEMBER]]
+            [ContentContainerControllerAccess::RULE_USER_GROUP_ONLY => [Space::USERGROUP_MEMBER, User::USERGROUP_SELF]]
         ];
     }